requirements.txt
logomaker
matplotlib
pandas
requests
streamlit
metapredict
biopython
pytest